requirements.txt
numpy
pyautogui
scikit-image
scipy
Cython
opencv-python-headless
Pillow
matplotlib
imgaug>=0.4
albumentations>0.5
graphviz
tqdm
pyyaml
easydict
torch>=1.4.0
torchvision>=0.5.0
tensorboard
future
cffi
ninja
rdp